Enable a 'url' arg to build_tests
The value of url overrides host, port and prefix if it is set.
This will make some live testing scenarios more straightforward.

Fixes #50
