+++ gabbi/runner.py
@@ -17,8 +17,6 @@ from importlib import import_module
 import sys
 import unittest
 
-from six.moves.urllib import parse as urlparse
-
 from gabbi import case
 from gabbi import handlers
 from gabbi.reporter import ConciseTestRunner
@@ -93,7 +91,7 @@ def run():
     )
 
     args = parser.parse_args()
-    host, port, prefix, force_ssl = process_target_args(
+    host, port, prefix, force_ssl = utils.host_info_from_target(
         args.target, args.prefix)
 
     # Initialize response handlers.
@@ -111,31 +109,6 @@ def run():
     result = ConciseTestRunner(
         verbosity=2, failfast=args.failfast).run(test_suite)
     sys.exit(not result.wasSuccessful())
-
-
-def process_target_args(target, prefix):
-    """Turn the argparse args into a host, port and prefix."""
-    force_ssl = False
-    split_url = urlparse.urlparse(target)
-
-    if split_url.scheme:
-        if split_url.scheme == 'https':
-            force_ssl = True
-        return split_url.hostname, split_url.port, split_url.path, force_ssl
-    else:
-        target = target
-        prefix = prefix
-
-    if ':' in target and '[' not in target:
-        host, port = target.rsplit(':', 1)
-    elif ']:' in target:
-        host, port = target.rsplit(':', 1)
-    else:
-        host = target
-        port = None
-    host = host.replace('[', '').replace(']', '')
-
-    return host, port, prefix, force_ssl
 
 
 def initialize_handlers(response_handlers):
