+++ gabbi/tests/test_runner.py
@@ -229,93 +229,6 @@ class RunnerTest(unittest.TestCase):
         self._stderr.write(sys.stderr.read())
 
 
-class RunnerHostArgParse(unittest.TestCase):
-
-    def _test_hostport(self, url_or_host, expected_host,
-                       provided_prefix=None, expected_port=None,
-                       expected_prefix=None, expected_ssl=False):
-        host, port, prefix, ssl = runner.process_target_args(
-            url_or_host, provided_prefix)
-
-        # normalize hosts, they are case insensitive
-        self.assertEqual(expected_host.lower(), host.lower())
-        # port can be a string or int depending on the inputs
-        self.assertEqual(expected_port, port)
-        self.assertEqual(expected_prefix, prefix)
-        self.assertEqual(expected_ssl, ssl)
-
-    def test_plain_url_no_port(self):
-        self._test_hostport('http://foobar.com/news',
-                            'foobar.com',
-                            expected_port=None,
-                            expected_prefix='/news')
-
-    def test_plain_url_with_port(self):
-        self._test_hostport('http://foobar.com:80/news',
-                            'foobar.com',
-                            expected_port=80,
-                            expected_prefix='/news')
-
-    def test_ssl_url(self):
-        self._test_hostport('https://foobar.com/news',
-                            'foobar.com',
-                            expected_prefix='/news',
-                            expected_ssl=True)
-
-    def test_ssl_port80_url(self):
-        self._test_hostport('https://foobar.com:80/news',
-                            'foobar.com',
-                            expected_prefix='/news',
-                            expected_port=80,
-                            expected_ssl=True)
-
-    def test_ssl_port_url(self):
-        self._test_hostport('https://foobar.com:999/news',
-                            'foobar.com',
-                            expected_prefix='/news',
-                            expected_port=999,
-                            expected_ssl=True)
-
-    def test_simple_hostport(self):
-        self._test_hostport('foobar.com:999',
-                            'foobar.com',
-                            expected_port='999')
-
-    def test_simple_hostport_with_prefix(self):
-        self._test_hostport('foobar.com:999',
-                            'foobar.com',
-                            provided_prefix='/news',
-                            expected_port='999',
-                            expected_prefix='/news')
-
-    def test_ipv6_url_long(self):
-        self._test_hostport(
-            'http://[FEDC:BA98:7654:3210:FEDC:BA98:7654:3210]:999/news',
-            'FEDC:BA98:7654:3210:FEDC:BA98:7654:3210',
-            expected_port=999,
-            expected_prefix='/news')
-
-    def test_ipv6_url_localhost(self):
-        self._test_hostport(
-            'http://[::1]:999/news',
-            '::1',
-            expected_port=999,
-            expected_prefix='/news')
-
-    def test_ipv6_host_localhost(self):
-        # If a user wants to use the hostport form, then they need
-        # to hack it with the brackets.
-        self._test_hostport(
-            '[::1]',
-            '::1')
-
-    def test_ipv6_hostport_localhost(self):
-        self._test_hostport(
-            '[::1]:999',
-            '::1',
-            expected_port='999')
-
-
 class HTMLResponseHandler(handlers.ResponseHandler):
 
     test_key_suffix = 'html'
